# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk platform validasi hasil kerja AI oleh ahli manusia.",
        "title": "AIValid API",
        "termsOfService": "https://aivalid.id/rules-content",
        "contact": {
            "name": "AIValid Support",
            "url": "https://aivalid.id/contact-support"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "api.aivalid.id",
    "basePath": "/api/v1",
    "paths": {
...
        "/validation-cases/{id}/workspace/apply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Apply as a validator for a validation case. Requires authentication.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Apply to validate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRepoTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/workspace/confidence/vote": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Vote confidence for a validator in a validation case workspace.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Vote confidence",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Validator user ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerVoteConfidenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRepoTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/workspace/consensus": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the current consensus status for a validation case workspace.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Get consensus status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerConsensusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/workspace/files": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Attach a file to the validation case workspace. Requires authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Upload workspace files",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "File details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerAttachRepoFileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRepoTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/workspace/finalize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Finalize the validation process for a case. Only the case owner can finalize.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Finalize validation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRepoTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/workspace/publish": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Publish the workspace for a validation case. Only the case owner can publish.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Publish workspace",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRepoTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/workspace/tree": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the file tree for a validation case workspace. Requires authentication.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Get workspace file tree",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRepoTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/workspace/validators/assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assign specific validators to a validation case. Only the case owner can assign.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Assign validators",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Validator IDs and panel size",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerAssignValidatorsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRepoTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/validation-cases/{id}/workspace/validators/auto-assign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Automatically assign validators to a validation case. Only the case owner can trigger.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ValidationCases-Workspace"
                ],
                "summary": "Auto-assign validators",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Validation Case ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Panel size (optional)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerAutoAssignValidatorsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerRepoTreeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.SwaggerAssignValidatorsRequest": {
            "type": "object",
            "properties": {
                "panel_size": {
                    "type": "integer",
                    "example": 3
                },
                "validator_user_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "handlers.SwaggerAttachRepoFileRequest": {
            "type": "object",
            "properties": {
                "document_id": {
                    "type": "string",
                    "example": "doc_abc123"
                },
                "kind": {
                    "type": "string",
                    "example": "evidence"
                },
                "label": {
                    "type": "string",
                    "example": "Screenshot bukti"
                },
                "visibility": {
                    "type": "string",
                    "example": "public"
                }
            }
        },
        "handlers.SwaggerAutoAssignValidatorsRequest": {
            "type": "object",
            "properties": {
                "panel_size": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
        "handlers.SwaggerConsensusResponse": {
            "type": "object",
            "properties": {
                "consensus": {}
            }
        },
...
        "handlers.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH001"
                },
                "details": {
                    "type": "string",
                    "example": "Detail tambahan"
                },
                "message": {
                    "type": "string",
                    "example": "Email atau password salah"
                }
            }
        },
...
        "handlers.SwaggerRepoTreeResponse": {
            "type": "object",
            "properties": {
                "repo_tree": {}
            }
        },
...
        "handlers.SwaggerVoteConfidenceRequest": {
            "type": "object",
            "properties": {
                "validator_user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin JWT token. Type \"Bearer\" followed by a space and admin JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}